Correccion de la funcion decode de los mendajes TFTP
Correccion de la escritura de caracteres.

Inicio modificacion de la lógica de confirmaciones en StopWaitProtocol

+++ src/practicas/bolqueII/tftp/datagram/headers/Header.java
@@ -29,7 +29,7 @@ public interface Header {
      */
     DatagramPacket encapsulate(InetAddress address, int port);
 
-    int getOpCode();
+    short getOpCode();
 
     String getFileName();
 }
